vionex-chatbot-service/requirements.txt
python-dotenv
grpcio
grpcio-tools